Fixing component key
